requirements.txt
audiosegment
docopt
matplotlib
numpy
pyAudio
pydub
pygame
pyzmq
scipy
tensorflow
scikit-learn
simpleaudio